fix nested mixin class
